+++ messages/set_super_passwords.go
@@ -4,10 +4,10 @@ import (
 	"github.com/uhppoted/uhppote-core/types"
 )
 
-// Request message to set the 'super' password assigned to a door.
+// Request message to set the 'super' passwords assigned to a door.
 // Passwords are in the range [0..999999], with 0 corresponding to
 // 'no password'
-type SetSuperPasswordRequest struct {
+type SetSuperPasswordsRequest struct {
 	MsgType      types.MsgType      `uhppote:"value:0x8c"` // message type (0x8c)
 	SerialNumber types.SerialNumber `uhppote:"offset:4"`   // controller serial number
 	Door         uint8              `uhppote:"offset:8"`   // door [1..4]
@@ -17,8 +17,8 @@ type SetSuperPasswordRequest struct {
 	Password4    uint32             `uhppote:"offset:24"`  // super password 4 [0..999999]
 }
 
-// Response message to a request set the 'super' password assigned to a door.
-type SetSuperPasswordResponse struct {
+// Response message to a request set the 'super' passwords assigned to a door.
+type SetSuperPasswordsResponse struct {
 	MsgType      types.MsgType      `uhppote:"value:0x8c"` // message type (0x8c)
 	SerialNumber types.SerialNumber `uhppote:"offset:4"`   // controller serial number
 	Succeeded    bool               `uhppote:"offset:8"`   // request succeeded/failed